packet: remove InsertPTS (#120)
this method is untested and unused, and the implementation is incorrect.

it was originally added in fcb2e8a5fcdf08bf45f14320b95b54443e9db658

but in 9a8460fcbdedb36ca30b33f8e59b56f12c218ab2 the usage was removed in favor
of gots.InsertPTS which has a different and more correct implementation.

as a library user, it is quite confusing to see two methods
in different packages with the same function signature and different
implementations. it seems best to remove this and force everyone to use
the tested version.